feat: isolate the whole report worker logic from service
